+++ src/dynamodb-actions/index.ts
@@ -3,14 +3,17 @@ import { v4 as uuid } from 'uuid';
 
 const dynamoDB = new AWS.DynamoDB.DocumentClient();
 
-/** put a to-do item in the db table */
-export function saveItemInDB(item: string, complete: boolean) {
+/** create a to-do item in the db table */
+export function createItemInDB(todoItem: string, complete: boolean) {
+  const timestamp = new Date().getTime();
   const params = {
-    TableName: "to-do-list",
+    TableName: process.env.TABLE_NAME,
     Item: {
       id: uuid(),
-      item,
-      complete
+      todoItem,
+      complete,
+      createdAt: timestamp,
+      updatedAt: timestamp
     }
   };
 
@@ -21,13 +24,32 @@ export function saveItemInDB(item: string, complete: boolean) {
     .catch(err => err);
 }
 
+/** update a to-do item in the db table */
+export function updateItemInDB(id: string, todoItem: string, complete: boolean) {
+  const timestamp = new Date().getTime();
+  const params = {
+    TableName: process.env.TABLE_NAME,
+    Key: { id },
+    ExpressionAttributeValues: {
+      ':todoItem': todoItem,
+      ':complete': complete,
+      ':updatedAt': timestamp,
+    },
+    UpdateExpression: 'SET todoItem = :todoItem, complete = :complete, updatedAt = :updatedAt',
+  };
+  console.log('updated params >>', params);
+  return dynamoDB
+    .update(params)
+    .promise()
+    .then(res => res)
+    .catch(err => err);
+} 
+
 /** get a to-do item from the db table */
 export function getItemFromDB(id: string) {
   const params = {
-    TableName: "to-do-list",
-    Key: {
-      id
-    }
+    TableName: process.env.TABLE_NAME,
+    Key: {id}
   };
 
   return dynamoDB
@@ -37,10 +59,23 @@ export function getItemFromDB(id: string) {
     .catch(err => err);
 }
 
+/** get all the to-do items from the db table */
+export function getAllItemsFromDB() {
+  const params = {
+    TableName: process.env.TABLE_NAME
+  };
+
+  return dynamoDB
+    .scan(params)
+    .promise()
+    .then(res => res)
+    .catch(err => err);
+}
+
 /** delete a to-do item from the db table */
 export function deleteItemFromDB(id: string) {
   const params = {
-    TableName: "to-do-list",
+    TableName: process.env.TABLE_NAME,
     Key: {
       id
     }