+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+<a name="v0.0.2"></a>
+## 0.0.2 (2021-01-05)
+
+#### Bug Fixes
+
+* NA
+
+#### Features
+
+* Read the environmental variables during build time using [serverless-dotenv-plugin](https://github.com/colynb/serverless-dotenv-plugin#readme) instead of using `dotenv` npm package.
+* Added update, delete and getAllToDoItems operations.
+
+#### Breaking Changes
+
+* NA
+
 
 <a name="v0.0.1"></a>
 ## 0.0.1 (2020-12-30)